# Speed up `QM.add_linear_from()`