--- a/asv.conf.json
+++ b/asv.conf.json
@@ -7,7 +7,7 @@
         "python -m pip install -r requirements.txt",
         "python -m pip install parameterized",
         "python setup.py build",
-        "PIP_NO_BUILD_ISOLATION=false python -mpip wheel --no-deps --no-index -w {build_cache_dir} {build_dir}"
+        "DIMOD_NUM_BUILD_JOBS=20 PIP_NO_BUILD_ISOLATION=false python -mpip wheel --no-deps --no-index -w {build_cache_dir} {build_dir}"
     ],
     "branches": ["main"],
     "environment_type": "virtualenv",