77fd6af995c2b940dedc4975d320c690aa019b31 (see https://github.com/dwavesystems/dimod/pull/1170) introduced a performance regression. This commit fixes it.

Comparing 0.10.17 to 0.11.0
```
+        311±10μs         646±30μs     2.08  qm.TimeAddLinearFrom.time_add_linear_from(<class 'int'>, <class 'dict'>)
+        322±20μs         630±20μs     1.95  qm.TimeAddLinearFrom.time_add_linear_from(<class 'int'>, <class 'list'>)
      1.07±0.04ms       1.69±0.2ms    ~1.57  qm.TimeAddLinearFrom.time_add_linear_from(<class 'str'>, <class 'dict'>)
+     1.06±0.03ms      1.41±0.04ms     1.32  qm.TimeAddLinearFrom.time_add_linear_from(<class 'str'>, <class 'list'>)
```
Comparing 0.10.17 to this commit
```
         311±10μs        295±0.8μs     0.95  qm.TimeAddLinearFrom.time_add_linear_from(<class 'int'>, <class 'dict'>)
         322±20μs          290±4μs    ~0.90  qm.TimeAddLinearFrom.time_add_linear_from(<class 'int'>, <class 'list'>)
-     1.07±0.04ms         971±10μs     0.90  qm.TimeAddLinearFrom.time_add_linear_from(<class 'str'>, <class 'dict'>)
      1.06±0.03ms         974±20μs     0.91  qm.TimeAddLinearFrom.time_add_linear_from(<class 'str'>, <class 'list'>)
```

there are other similar patterns we could do for `add_quadratic_from()` but in that case the binary search for insertion dominates for moderately sized models